+++ lib/spout/helpers/subject_loader.rb
@@ -5,7 +5,7 @@ module Spout
   module Helpers
     class SubjectLoader
       attr_accessor :subjects
-      attr_reader :all_methods
+      attr_reader :all_methods, :all_domains
 
       def initialize(variable_files, valid_ids, standard_version, number_of_rows, visit)
         @subjects = []
@@ -15,6 +15,7 @@ module Spout
         @number_of_rows = number_of_rows
         @visit = visit
         @all_methods = {}
+        @all_domains = []
       end
 
       def load_subjects_from_csvs!
@@ -51,7 +52,6 @@ module Spout
               end
             end
             # puts "Memory Used: " + (`ps -o rss -p #{$$}`.strip.split.last.to_i / 1024).to_s + " MB" if count % 1000 == 0
-            # break if count >= 1000
             break if @number_of_rows != nil and count >= @number_of_rows
           end
         end
@@ -69,6 +69,17 @@ module Spout
           @subjects.each{ |s| s.send(method) != nil ? s.send("#{method}=", s.send("#{method}").to_f) : nil }
         end
         @subjects
+      end
+
+      def load_variable_domains!
+        @variable_files.each do |variable_file|
+          json = JSON.parse(File.read(variable_file)) rescue json = nil
+          next unless json
+          next unless ["choices"].include?(json["type"])
+          domain = json['domain'].to_s.downcase
+          @all_domains << domain
+        end
+        @all_domains = @all_domains.compact.uniq.sort
       end
     end
   end